+++ common/src/main/java/dev/toma/configuration/client/screen/ConfigScreen.java
@@ -73,18 +73,20 @@ public class ConfigScreen extends AbstractConfigScreen {
 
     @Override
     public void render(GuiGraphics graphics, int mouseX, int mouseY, float partialTicks) {
-        renderBackground(graphics, mouseX, mouseY, partialTicks);
-        // HEADER
-        ConfigTheme.Header themeHeader = this.theme.getHeader();
-        ConfigTheme.Footer footer = this.theme.getFooter();
-        Component headerLabel = themeHeader.customText() != null ? themeHeader.customText() : this.title;
-        int titleWidth = this.font.width(headerLabel);
-        graphics.drawString(font, headerLabel, (this.width - titleWidth) / 2, (HEADER_HEIGHT - this.font.lineHeight) / 2, themeHeader.foregroundColor(), true);
-        graphics.fill(0, 0, width, HEADER_HEIGHT, themeHeader.backgroundColor());
-        graphics.fill(0, height - FOOTER_HEIGHT, width, height, footer.backgroundColor());
-        Integer fillColor = this.theme.getBackgroundFillColor();
-        if (fillColor != null) {
-            graphics.fill(0, HEADER_HEIGHT, width, height - FOOTER_HEIGHT, fillColor);
+        if (canRenderBackground(minecraft)) {
+            renderBackground(graphics, mouseX, mouseY, partialTicks);
+            // HEADER
+            ConfigTheme.Header themeHeader = this.theme.getHeader();
+            ConfigTheme.Footer footer = this.theme.getFooter();
+            Component headerLabel = themeHeader.customText() != null ? themeHeader.customText() : this.title;
+            int titleWidth = this.font.width(headerLabel);
+            graphics.drawString(font, headerLabel, (this.width - titleWidth) / 2, (HEADER_HEIGHT - this.font.lineHeight) / 2, themeHeader.foregroundColor(), true);
+            graphics.fill(0, 0, width, HEADER_HEIGHT, themeHeader.backgroundColor());
+            graphics.fill(0, height - FOOTER_HEIGHT, width, height, footer.backgroundColor());
+            Integer fillColor = this.theme.getBackgroundFillColor();
+            if (fillColor != null) {
+                graphics.fill(0, HEADER_HEIGHT, width, height - FOOTER_HEIGHT, fillColor);
+            }
         }
         renderables.forEach(renderable -> renderable.render(graphics, mouseX, mouseY, partialTicks));
         ConfigTheme.Scrollbar scrollbar = this.theme.getScrollbar();