+++ common/src/main/java/dev/toma/configuration/config/adapter/TypeAdapter.java
@@ -37,11 +37,11 @@ public abstract class TypeAdapter<V> {
     }
 
     public record TypeAttributes<V>(String configOwner, String id, V value, TypeAdapter.AdapterContext context,
-                                    Configurable.LocalizationKey localization, String[] fileComments, boolean localizeComments
+                                    Configurable.LocalizationKey localization, String translationKey, String[] fileComments, boolean localizeComments
     ) {
 
         public <R> TypeAttributes<R> child(String id, R value, TypeAdapter.AdapterContext ctx) {
-            return new TypeAttributes<>(configOwner, id, value, ctx, localization, fileComments, localizeComments);
+            return new TypeAttributes<>(configOwner, id, value, ctx, localization, translationKey, fileComments, localizeComments);
         }
     }
 }