+++ CHANGELOG.md
@@ -1,3 +1,3 @@
-# 1.21.5-3.1.2
-- Renamed "Synchronized" attribute to more user-friendly text
-- Fixed issue with broken navigation in config selection UI
+# 1.21.5-3.2.0
+- Added `value` attribute to `@Configurable` annotation - can be used to set custom translation keys
+- Added `hideBackground` setting - will disable configuration background screen when playing. Can be useful for HUD element configurations